feat: toolbox image manifest
Not yet being built with GitHub Action, I will have to take care of this later. It is intended to workaround some tooling not being available everywhere, which is a pain most of the time. This container might get a bit larger but it will not (in nearer future) cause even nearly equally pain as not having my tooling with me.